requirements.txt
flask==3.1.1
twilio==9.4.1
openai==1.78.0
python-dotenv==1.1.0
apscheduler==3.11.0
mysql-connector-python==9.4.0
gunicorn==23.0.0



